update: 更新接口规范

+++ src/main/java/edu/tust/neusoft/backend/repository/WalletRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface WalletRepository extends JpaRepository<Wallet, Integer> {
-    Optional<Wallet> findByUserId(int userId);
+    Optional<Wallet> findByUserId(long userId);
 }
